+++ Recolor/Systems/ColorVariations/CustomColorVariationSystem.cs
@@ -177,9 +177,17 @@ namespace Recolor.Systems.ColorVariations
             NativeArray<Entity> entities = prefabRefQuery.ToEntityArray(Allocator.Temp);
             foreach (Entity e in entities)
             {
-                if (EntityManager.TryGetComponent(e, out PrefabRef currentPrefabRef) && EntityManager.TryGetBuffer(currentPrefabRef.m_Prefab, isReadOnly: true, out DynamicBuffer<SubMesh> currentSubMeshBuffer) && prefabsNeedingUpdates.Contains(currentSubMeshBuffer[0].m_SubMesh))
+                if (EntityManager.TryGetComponent(e, out PrefabRef currentPrefabRef) &&
+                    EntityManager.TryGetBuffer(currentPrefabRef.m_Prefab, isReadOnly: true, out DynamicBuffer<SubMesh> currentSubMeshBuffer))
                 {
-                    buffer.AddComponent<BatchesUpdated>(e);
+                    for (int i = 0; i < currentSubMeshBuffer.Length; i++)
+                    {
+                        if (prefabsNeedingUpdates.Contains(currentSubMeshBuffer[i].m_SubMesh))
+                        {
+                            buffer.AddComponent<BatchesUpdated>(e);
+                            break;
+                        }
+                    }
                 }
             }
         }