+++ Recolor/Systems/SelectedInfoPanel/SIPColorFieldsSystem.Initialization.cs
@@ -247,83 +247,8 @@ namespace Recolor.Systems.SelectedInfoPanel
                 Mod.Instance.Settings.ShowHexaDecimals = m_ShowHexaDecimals.Value;
                 Mod.Instance.Settings.ApplyAndSave();
             });
-            CreateTrigger("ReduceSubMeshIndex", () =>
-            {
-                if (m_SubMeshData.Value.SingleSubMesh == ButtonState.On &&
-                    m_SubMeshData.Value.SubMeshIndex > 0)
-                {
-                    m_SubMeshData.Value.SubMeshIndex = Mathf.Clamp(m_SubMeshData.Value.SubMeshIndex - 1, 0, m_SubMeshData.Value.SubMeshLength - 1);
-                }
-                else if (m_SubMeshData.Value.SingleSubMesh == ButtonState.On &&
-                         m_SubMeshData.Value.SubMeshIndex <= 0)
-                {
-                    m_SubMeshData.Value.SubMeshIndex = m_SubMeshData.Value.SubMeshLength - 1;
-                }
-                else if (m_SubMeshData.Value.MatchingSubMeshes == ButtonState.On &&
-                         EntityManager.TryGetBuffer(m_CurrentPrefabEntity, isReadOnly: true, out DynamicBuffer<SubMesh> submeshes))
-                {
-                    int attempts = 0;
-                    int index = m_SubMeshData.Value.SubMeshIndex;
-                    while (attempts < submeshes.Length)
-                    {
-                        index--;
-                        if (index < 0)
-                        {
-                            index = submeshes.Length - 1;
-                        }
-
-                        if (m_PrefabSystem.GetPrefabName(submeshes[index].m_SubMesh) != m_SubMeshData.Value.SubMeshName)
-                        {
-                            m_SubMeshData.Value.SubMeshIndex = index;
-                            break;
-                        }
-
-                        attempts++;
-                    }
-                }
-
-                m_SubMeshData.Binding.TriggerUpdate();
-                m_PreviouslySelectedEntity = Entity.Null;
-            });
-
-            CreateTrigger("IncreaseSubMeshIndex", () =>
-            {
-                if (m_SubMeshData.Value.SingleSubMesh == ButtonState.On &&
-                    m_SubMeshData.Value.SubMeshIndex < m_SubMeshData.Value.SubMeshLength - 1)
-                {
-                    m_SubMeshData.Value.SubMeshIndex = Mathf.Clamp(m_SubMeshData.Value.SubMeshIndex + 1, 0, m_SubMeshData.Value.SubMeshLength - 1);
-                }
-                else if (m_SubMeshData.Value.SingleSubMesh == ButtonState.On &&
-                         m_SubMeshData.Value.SubMeshIndex >= m_SubMeshData.Value.SubMeshLength - 1)
-                {
-                    m_SubMeshData.Value.SubMeshIndex = 0;
-                }
-                else if (m_SubMeshData.Value.MatchingSubMeshes == ButtonState.On &&
-                         EntityManager.TryGetBuffer(m_CurrentPrefabEntity, isReadOnly: true, out DynamicBuffer<SubMesh> submeshes))
-                {
-                    int attempts = 0;
-                    int index = m_SubMeshData.Value.SubMeshIndex;
-                    while (attempts < submeshes.Length)
-                    {
-                        index++;
-                        if (index >= submeshes.Length)
-                        {
-                            index = 0;
-                        }
-
-                        if (m_PrefabSystem.GetPrefabName(submeshes[index].m_SubMesh) != m_SubMeshData.Value.SubMeshName)
-                        {
-                            m_SubMeshData.Value.SubMeshIndex = index;
-                            break;
-                        }
-
-                        attempts++;
-                    }
-                }
-
-                m_SubMeshData.Binding.TriggerUpdate();
-                m_PreviouslySelectedEntity = Entity.Null;
-            });
+            CreateTrigger("ReduceSubMeshIndex", ReduceSubMeshIndex);
+            CreateTrigger("IncreaseSubMeshIndex", IncreaseSubMeshIndex);
 
             CreateTrigger("ChangeSubMeshScope", (int newScope) =>
             {